Split up API tests
